geocoding api.

+++ db/migrate/20150406185313_create_markets.rb
@@ -4,8 +4,8 @@ class CreateMarkets < ActiveRecord::Migration
       t.string :name
       t.string :address
       t.string :zipcode
-      t.integer :longitude
-      t.integer :latitude
+      t.float :longitude
+      t.float :latitude
       t.boolean :accepts_ebt
       t.belongs_to :city
     end
